feat(pkg/util, pkg/component): 添加了一些基础组件，完成了缓存组件开发及其测试文件编写，部分完成翻页组件 (#87)
* fix(app/businessOrder)修改了order的部分字段,部分重写了businessOrder

* fix(app/businessOrder)修改了order的部分字段,重写了businessOrder

* refactor(app/address)增加了redis缓存部分,缓存的序列化和反序列化都采用json格式

* fix(app/address)修复了redis缓存,增加了logrus作为日志输出

* refactor(app/businessOrder)为businessOrder增加了缓存,采用的方案为id列表+详情分级缓存,后续应该将order部分和product部分也如此处理

* refactor(app/businessOrder)统一了businessOrder,order和address的缓存策略

* refactor(app/...)统一了ID字段为uint64类型,新增了角色表和权限表以及对应的关联表

* refactor(app/...)重新规划了位置,错误被放入exception下,辅助函数被放入helper下

* feat(app/users)增加了方法,补充了网关的infra部分

* feat(apps/pkg)添加了component组件,内部包含了一个深分页查询组件,采用游标查询,支持前后翻页,以及一个缓存组件

* fix(apps/pkg)增加了一个conditionBuilder

* fix(apps/test)清理了test

* feat(apps/users)实现了部分user接口

* fix(apps/cache)修复了cache

* feat(pkg/util): 添加缓存组件 cache 的单元测试

* fix(apps/ai): 修复订单服务客户端创建报错
fix(apps/product): 更新GetCategoryFromProduct方法以支持缓存上下文
fix(pkg/els): 修改DeleteProduct测试中的id类型为uint64
fix(pkg/util): 更新TestUploadImages中的id类型为uint64

* fix(go.mod): 移除不必要的依赖项并修复依赖顺序

* docs(apps/cache)增加了cache的注释说明和component组件的缓存说明

---------

+++ apps/api/handlers/address/address_update.go
@@ -11,7 +11,7 @@ import (
 )
 
 func Update(ctx context.Context, c *app.RequestContext) {
-	userId, ok := ctx.Value("userId").(uint32)
+	userId, ok := ctx.Value("userId").(uint64)
 	if !ok {
 		c.JSON(consts.StatusBadRequest, utils.H{
 			"error": "userId must be a number",
