+++ apps/api/handlers/shop/shop_add.go
@@ -13,7 +13,7 @@ import (
 )
 
 func Add(ctx context.Context, c *app.RequestContext) {
-	userId, ok := ctx.Value("userId").(uint32)
+	userId, ok := ctx.Value("userId").(uint64)
 	if !ok {
 		c.JSON(consts.StatusBadRequest, utils.H{
 			"error": "userId must be a number",
@@ -21,7 +21,7 @@ func Add(ctx context.Context, c *app.RequestContext) {
 		return
 	}
 	type ProductInfo struct {
-		ShopID      uint32   `json:"shop_id"`
+		ShopID      uint64   `json:"shop_id"`
 		Name        string   `json:"product_name"`
 		Description string   `json:"product_description"`
 		Picture     string   `json:"product_picture"`
