+++ apps/api/handlers/user/user_info.go
@@ -13,7 +13,7 @@ import (
 
 func GetInfo(ctx context.Context, c *app.RequestContext) {
 	//获取用户id
-	userId, ok := ctx.Value("userId").(uint32)
+	userId, ok := ctx.Value("userId").(uint64)
 	if !ok {
 		c.JSON(consts.StatusBadRequest, utils.H{
 			"error": "userId must be a number",
