+++ pkg/util/oss_test.go
@@ -5,7 +5,7 @@ import "testing"
 func TestUploadImages(t *testing.T) {
 	localFilePath := "img.png"
 	serviceName := "test"
-	id := uint32(1)
+	id := uint64(1)
 	url, err := UploadImages(localFilePath, serviceName, id)
 	if err != nil {
 		t.Errorf("UploadImages() error = %v", err)
